backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
prisma==0.13.1
asyncpg==0.29.0
httpx==0.27.0
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
pydantic==2.7.1
pytest==8.2.0
